+++ viscy/representation/evalutation/clustering.py
@@ -7,6 +7,76 @@ from sklearn.metrics import (
     normalized_mutual_info_score,
 )
 from sklearn.neighbors import KNeighborsClassifier
+from sklearn.mixture import GaussianMixture
+import numpy as np
+
+class GMMClustering:
+    def __init__(self, features_data, n_clusters_range=np.arange(2, 10)):
+        self.features_data = features_data
+        self.n_clusters_range = n_clusters_range
+        self.best_n_clusters = None
+        self.best_gmm = None
+        self.aic_scores = None
+        self.bic_scores = None
+
+    def find_best_n_clusters(self):
+        """Find the best number of clusters using AIC/BIC scores."""
+        aic_scores = []
+        bic_scores = []
+        for n in self.n_clusters_range:
+            gmm = GaussianMixture(n_components=n, random_state=42)
+            gmm.fit(self.features_data)
+            aic_scores.append(gmm.aic(self.features_data))
+            bic_scores.append(gmm.bic(self.features_data))
+
+        self.aic_scores = aic_scores
+        self.bic_scores = bic_scores
+
+        return aic_scores, bic_scores
+
+    def fit_best_model(self, criterion="bic", n_clusters=None):
+        """
+        Fit the best GMM model based on AIC or BIC scores, or a user-specified number of clusters.
+
+        Parameters:
+        - criterion: 'aic' or 'bic' to select the best model based on the chosen criterion.
+        - n_clusters: Specify a fixed number of clusters (overrides the 'best' search).
+        """
+        # Case 1: If the user provides n_clusters, use it directly
+        if n_clusters is not None:
+            self.best_n_clusters = n_clusters
+
+        # Case 2: If no n_clusters is provided but find_best_n_clusters was run, use stored AIC/BIC results
+        elif self.aic_scores is not None and self.bic_scores is not None:
+            if criterion == "bic":
+                self.best_n_clusters = self.n_clusters_range[np.argmin(self.bic_scores)]
+            else:
+                self.best_n_clusters = self.n_clusters_range[np.argmin(self.aic_scores)]
+
+        # Case 3: If find_best_n_clusters hasn't been run, compute AIC/BIC scores now
+        else:
+            aic_scores, bic_scores = self.find_best_n_clusters()
+            if criterion == "bic":
+                self.best_n_clusters = self.n_clusters_range[np.argmin(bic_scores)]
+            else:
+                self.best_n_clusters = self.n_clusters_range[np.argmin(aic_scores)]
+
+        self.best_gmm = GaussianMixture(
+            n_components=self.best_n_clusters, random_state=42
+        )
+        self.best_gmm.fit(self.features_data)
+
+        return self.best_gmm
+
+    def predict_clusters(self):
+        """Run prediction on the fitted best GMM model."""
+        if self.best_gmm is None:
+            raise Exception(
+                "No GMM model is fitted yet. Please run fit_best_model() first."
+            )
+        cluster_labels = self.best_gmm.predict(self.features_data)
+        return cluster_labels
+
 
 
 def knn_accuracy(embeddings, annotations, k=5):
